Transfer the logic of the app to the game component
